100-prime_factor.c

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	long i, max;
-	int j = 612852475143;
+	long j = 612852475143;
 	double k = sqrt(j);
 
 		for (i = 1; i <= k; i++)
